[SYCL] Fix kernel names on fpga
