+++ clang/lib/Sema/SemaSYCL.cpp
@@ -28,6 +28,7 @@
 #include "llvm/Support/FileSystem.h"
 #include "llvm/Support/Path.h"
 #include "llvm/Support/raw_ostream.h"
+#include "llvm/Support/SHA1.h"
 
 #include <array>
 #include <functional>
@@ -1087,10 +1088,8 @@ constructKernelName(Sema &S, const FunctionDecl *KernelCallerFunc,
 
   MC.mangleTypeName(KernelNameType, Out);
 
-  std::string Res = std::string(Out.str());
-  std::string Str =
-      SYCLUniqueStableNameExpr::ComputeName(S.getASTContext(), KernelNameType);
-  return {Res, Str};
+  return {std::string(Out.str()), SYCLUniqueStableNameExpr::ComputeName(
+                                      S.getASTContext(), KernelNameType)};
 }
 
 static bool isDefaultSPIRArch(ASTContext &Context) {
@@ -3927,6 +3926,82 @@ void Sema::copySYCLKernelAttrs(const CXXRecordDecl *KernelObj) {
   }
 }
 
+/// Compute a unique name that is consumable by sycl_vxx
+std::string computeUniqueSYCLVXXName(StringRef Demangle) {
+  /// VXX has a maximum of 64 character for the name of the kernel function
+  /// plus the name of one parameter.
+  /// Those characters need to be used wisely to prevent name collisions.
+  /// It is also useful to use a name that is understandable by the user,
+  /// so we add only 8 character of hash and only if needed.
+  /// The first character cannot be an underscore or a digit.
+  /// An underscore can only be found between two non underscore character
+  /// (No __ or undersocre at end or start of name)
+  constexpr unsigned MaxVXXSize = 30;
+  /// Some transformations might make 2 kernel identifiers the same.
+  /// Allow adding a hash when such transformations are made to avoid possible
+  /// name conflict.
+  bool ForceHash = false;
+
+  std::string Result;
+  Result.reserve(Demangle.size());
+
+  for (char c : Demangle) {
+    if (!isAlphanumeric(c)) {
+      // Do not repeat _ in the cleaned-up name
+      if (!Result.empty() && Result.back() == '_')
+        continue;
+      c = '_';
+    }
+    Result.push_back(c);
+  }
+
+  // Replace first kernel character name by a 'k' to be compatible with SPIR
+  if ((Result.front() == '_' || isDigit(Result.front()))) {
+    Result.front() = 'k';
+    ForceHash = true;
+  }
+
+  // Vivado IP naming requirements forbids having '_' as a last character
+  if (Result.back() == '_') {
+    ForceHash = true;
+  }
+
+  /// The name alone is guaranteed to be unique, so if fits in the size, it is
+  /// enough.
+  if (Result.size() < MaxVXXSize && !ForceHash) {
+    return Result;
+  }
+
+  if (Result.size() > (MaxVXXSize - 9)) {
+    /// 9 for 8 characters of hash and an '_'.
+    Result.erase(0, Result.size() - (MaxVXXSize - 9));
+  }
+
+  if (Result.front() == '_')
+    Result.front() = 'u';
+  if(isDigit(Result.front()))
+    Result.front() = 'a' + Result.front() - '0';
+
+  if (Result.back() != '_')
+    Result.push_back('_');
+
+  /// Sadly there is only 63 valid characters in C identifiers and v++ doesn't
+  /// deal well with double underscores in identifiers. So A and B are
+  /// repeated. This doesn't hurt entropy too much because it is just 2 out
+  /// of 64.
+  Result += llvm::SHA1::hashToString(
+      llvm::ArrayRef<uint8_t>{reinterpret_cast<const uint8_t *>(Demangle.data()),
+                              Demangle.size()},
+      "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
+      "abcdefghijklmnopqrstuvwxyz"
+      "0123456789AB");
+
+  if (Result.size() > MaxVXXSize)
+    Result.resize(MaxVXXSize);
+
+  return Result;
+}
+
 void Sema::SetSYCLKernelNames() {
   std::unique_ptr<MangleContext> MangleCtx(
       getASTContext().createMangleContext());
@@ -3939,6 +4014,8 @@ void Sema::SetSYCLKernelNames() {
         constructKernelName(*this, Pair.first, *MangleCtx);
     StringRef KernelName(
         IsSYCLUnnamedKernel(*this, Pair.first) ? StableName : CalculatedName);
+    std::string LastingName = computeUniqueSYCLVXXName(KernelName);
+    KernelName = LastingName;
 
     getSyclIntegrationHeader().updateKernelNames(Pair.first, KernelName,
                                                  StableName);
